refactor: extract app dialogs from main gui
